core: Spoof build properties
SafetyNet's CTS profile attestation checks whether Build.MODEL matches
that of the device's stock OS and checking whether it's a device that
matched the CTS Profile.

Spoof the model by setting Build.MODEL to "Pixel 4a (5G)" for
Google Play Services to help pass SafetyNet.

Also while on it, add a model check for Genshin Impact internal testing
tasks that we conduct by setting the Build.MODEL to "Pixel 5 Pro" if
ro.boot.hardware is "gourami" to make sure it not clash with other
devices that's not supposed to be enrolled.

REF: https://github.com/ProtonAOSP/android_frameworks_base/commit/b8adfb5d711c0381350901a40a693b0a85fb9d76
BUG: h/gourami/labs/c/1539827478/6247
BUG: h/gourami/labs/c/1539827478/6259
BUG: h/hoyo/labs/int/a/187
[Raphiel] : Reworded commit description for our usecase
Change-Id: Ic762d8853967bf78db614efaf46166e527cb397a
Signed-off-by: Raphiel Rollerscaperers <rapherion@raphielgang.org>